Rename Packagist class to represent project
